% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCVS_fun.R
\name{BIC_TCVS}
\alias{BIC_TCVS}
\title{BIC Calculation for CVXR Optimization}
\usage{
BIC_TCVS(
  Z,
  Y,
  p,
  group,
  maxlam,
  minlam,
  nlam,
  lambda_seq,
  A0,
  CoefNormalization = T
)
}
\arguments{
\item{Z}{the design matrix matrix.}

\item{Y}{the response vector.}

\item{p}{the number of features the for the simulated data.}

\item{group}{the taxonomic structure of OTUs.}

\item{maxlam}{the maximum lambda value for the tuning parameter lambda candidates.}

\item{minlam}{the minimum lambda value for the tuning parameter lambda candidates.}

\item{nlam}{the number of lambda value candidates.}

\item{lambda_seq}{the sequence of lambda values. If NULL, a sequence will be generated.}

\item{A0}{the constraint matrix.}

\item{CoefNormalization}{Logical. Whether to normalize the coefficients within each group. The default is T.}
}
\value{
A list containing the following elements:
\itemize{
  \item total.results - A list of results for each lambda value.
  \item best_lambda_BIC - The lambda value with the minimum BIC.
}
}
\description{
This function calculates the Bayesian Information Criterion (BIC) for a given set of parameters using the CVXR package.
}
