% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCVS_fun.R
\name{cvxr.result}
\alias{cvxr.result}
\title{CVXR Optimization Result}
\usage{
cvxr.result(y, Z, A0, lambda, p, group, CoefNormalization = T)
}
\arguments{
\item{y}{the response vector.}

\item{Z}{the design matrix matrix.}

\item{A0}{The constraint matrix, which enforces the condition that the coefficients of the original features and their
knockoff counterparts sum to zero, respectively.}

\item{lambda}{the regularization parameter.}

\item{p}{the number of features the for the simulated compositional data.}

\item{group}{the taxonomic structure of OTUs.}

\item{CoefNormalization}{Logical. Whether to normalize the coefficients within each group. The default is T.}
}
\value{
The estimated coefficients from the optimization.
}
\description{
This function performs an optimization using the CVXR package to solve an augmented regression problem with the penalty.
}
