% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\docType{data}
\name{dd}
\alias{dd}
\title{Example Microbiome Dataset}
\format{
An object of class `list`.
}
\usage{
data(DirMultOutput)
}
\description{
A dataset containing the counts of 856 taxa from an upper-respiratory-tract microbiome study.
}
\keyword{datasets}
