% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCVS_fun.R
\name{get.all.knockoff}
\alias{get.all.knockoff}
\title{Generate Knockoff Copy Matrix}
\usage{
get.all.knockoff(Z, seed = 2023)
}
\arguments{
\item{Z}{the original matrix.}

\item{seed}{the seed for random number generation. The default is 2023.}
}
\value{
The knockoff copy matrix.
}
\description{
This function generates the knockoff copies using a function from knockoff.
}
