% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCVS_fun.R
\name{get.all.OTU}
\alias{get.all.OTU}
\title{Generate OTU Matrix}
\usage{
get.all.OTU(
  n,
  p,
  trans = c(0, 1, 2),
  type,
  normalizeMethod,
  pseudocount,
  seed = 2023
)
}
\arguments{
\item{n}{the number of samples for the simulated compositional data.}

\item{p}{the number of features the for the simulated compositional data.}

\item{trans}{an integer specifying the type of transformation applied to the data.
The options are 0 (no transformation), 1 (log transformation), and 2 (clr transformation).}

\item{type}{the type of the simulated data to be generated. The options are "Lognormal_previous" and "Dirmult".}

\item{normalizeMethod}{the normalization method.}

\item{pseudocount}{the pseudocount value.}

\item{seed}{an integer value used to set the seed of the random number generator for ensuring reproducibility.
The default is 2023.}
}
\value{
The generated OTU matrix.
}
\description{
This function generates the OTU matrix.
}
