% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCVS_fun.R
\name{get.all.y}
\alias{get.all.y}
\title{Generate Response Variable Based on the CLR matrix Z}
\usage{
get.all.y(Z, setting = c(1, 2), beta.slack.factor = 1, seed = 2023)
}
\arguments{
\item{Z}{the CLR transforamtion matrix (design matrix).}

\item{setting}{a numeric vector indicating the setting for coefficient generation.
Setting 1 generates a fixed set of coefficients with specific non-zero values.
Setting 2 generates coefficients by sampling from a uniform distribution.}

\item{beta.slack.factor}{an optional numeric factor to adjust the coefficients by a given factor.
If not provided, coefficients are used as defined by the `setting`.
The default is 1.}

\item{seed}{an integer value to set the seed for random number generation to ensure reproducibility.}
}
\value{
A numeric vector representing the generated response variable `y`.
}
\description{
This function generates a response variable `y` as a linear combination of predictors
specified in matrix `Z`, influenced by a vector of coefficients `beta`. The coefficients
are determined based on predefined settings and optionally modified by a slack factor.
Random noise is added to the response to simulate more realistic scenarios.
}
\examples{
  Z <- matrix(rnorm(100 * 20), ncol = 20)
  y <- get.all.y(Z, setting = 1, beta.slack.factor = 1, seed = 2023)
}
