% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCVS_fun.R
\name{Group.lasso}
\alias{Group.lasso}
\title{Group Lasso Penalty with L2 Norm (Without Square)}
\usage{
Group.lasso(beta, group, lambda, p, CoefNormalization = T)
}
\arguments{
\item{beta}{a numeric vector of coefficients.}

\item{group}{the taxonomic structure of OTUs.}

\item{lambda}{the regularization parameter.}

\item{p}{the number of features the for the simulated compositional data.}

\item{CoefNormalization}{Logical. Whether to normalize the coefficients within each group. The default is T.}
}
\value{
The computed penalty.
}
\description{
This function computes the penalty term in the augmented problem incorporating auxiliary knockoff copies and tree structure with
L2 norm (without square) for a given set of coefficients, groups, and a regularization parameter.
}
