% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCVS_fun.R
\name{TCVS}
\alias{TCVS}
\title{A Tree-guided Compositional Variable Selection method (TCVS)}
\usage{
TCVS(X, Z, y, P, method = "BIC", maxlam, minlam, nlam, q = 0.05, seed = 2023)
}
\arguments{
\item{X}{the compositional covariate matrix.}

\item{Z}{the CLR transforamtion matrix (design matrix).}

\item{y}{the response vector.}

\item{P}{the taxonomic structure of OTUs.}

\item{method}{the method used for selecting the tuning parameter. The default is "BIC".}

\item{maxlam}{the maximum lambda value for the tuning parameter lambda candidates.}

\item{minlam}{the minimum lambda value for the tuning parameter lambda candidates.}

\item{nlam}{the number of lambda value candidates.}

\item{q}{A threshold value between 0 and 1. The default is 0.05.}

\item{seed}{an integer value used to set the seed of the random number generator for ensuring reproducibility.
The default is 2023.}
}
\value{
A list containing the following elements:
\itemize{
  \item TCVS.beta.hat - The estimated coefficients from the TCVS method.
  \item S.TCVS - The selected variables from the TCVS method.
}
}
\description{
This function allows you to
identify outcome-associated components (OTUs) within high-dimensional microbial compositional data with
hierarchical taxonomic tree information inherent among microbial taxa.
}
\examples{
# Example usage
library(GUniFrac)
n = 200
p = 60
sim.setting = 1
method = "BIC"
seed = 2023
type = "Dirmult" # type to generate compositional data
normalizeMethod = "Rowsum"
beta.slack.factor = NULL
pseudocount = 0.5
maxlam = 0.1
minlam = 1e-7
nlam = 2
q = 0.05
data(P_60, package = "TCVS") # P matrix
X <- get.all.OTU(n, p, trans = 0, type, normalizeMethod, pseudocount, seed = seed)
# clr transformation
Z <- get.all.OTU(n, p, trans = 2, type, normalizeMethod, pseudocount, seed = seed)
y <- get.all.y(Z, setting = sim.setting, beta.slack.factor, seed = seed)
result.TCVS <- TCVS(
  X = X,
  Z = Z,
  y = y,
  P = P,
  method = method,
  maxlam = maxlam,
  minlam = minlam,
  nlam = nlam,
  q = q,
  seed = seed
)

}
