% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCVS_fun.R
\name{TCVS_Generate_X_Z}
\alias{TCVS_Generate_X_Z}
\title{Generate compositional matrix X and CLR transformation matrix Z}
\usage{
TCVS_Generate_X_Z(
  n,
  p,
  type = c("Lognormal_previous", "Dirmult"),
  normalizeMethod = "Rowsum",
  pseudocount
)
}
\arguments{
\item{n}{the number of samples for the simulated compositional data.}

\item{p}{the number of features the for the simulated compositional data.}

\item{type}{the type of the simulated data to be generated. The options are "Lognormal_previous" and "Dirmult".}

\item{normalizeMethod}{the method of normalization when type = "Dirmult". The default is "Rowsum".}

\item{pseudocount}{a pseudo-count value added to the count matrix to prevent zero values in the read count.}
}
\value{
A list containing the following matrices:
\itemize{
  \item X - The compositional matrix.
  \item log_X - The logarithm of the compositional matrix.
  \item Z - The CLR transformation matrix.
}
}
\description{
This function generates the X and Z matrices based on the specified type and parameters.
}
\examples{
# Example usage for Dirmult type
result <- TCVS_Generate_X_Z(n = 200, p = 60, type="Lognormal_previous",
normalizeMethod, pseudocount = 0.5)

}
